# src/main/resources/META-INF/openapi.yaml
openapi: 3.0.1
info:
  title: Personal Finances V1
  version: 1.0.0
  description: |
    API para registrar, consultar, actualizar, eliminar, refrescar e importar/exportar gastos personales.

tags:
  - name: Expenses
  - name: CSV Expenses
  - name: Categories

paths:
  /poc/personal-finances/v1/expenses:
    get:
      tags:
        - Expenses
      summary: Buscar gastos con paginación por cursor
      description: |
        Buscar gastos del usuario aplicando filtros opcionales y paginación basada en cursor.
      operationId: searchExpenses
      parameters:
        - $ref: '#/components/parameters/CallerNameHeader'
        - $ref: '#/components/parameters/TraceParentHeader'
        - $ref: '#/components/parameters/UserCodeHeader'
        - $ref: '#/components/parameters/RecipientQuery'
        - $ref: '#/components/parameters/CategoryQuery'
        - $ref: '#/components/parameters/CurrencyQuery'
        - $ref: '#/components/parameters/FromDateQuery'
        - $ref: '#/components/parameters/ToDateQuery'
        - $ref: '#/components/parameters/EncodedCursorQuery'
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseSearchResponseDto'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Expenses
      summary: Registrar gasto personal
      description: Registrar gasto personal.
      operationId: saveExpense
      parameters:
        - $ref: '#/components/parameters/CallerNameHeader'
        - $ref: '#/components/parameters/TraceParentHeader'
        - $ref: '#/components/parameters/UserCodeHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseSaveRequestDto'
      responses:
        '201':
          description: Created
          headers:
            Location:
              description: URI relativa del recurso creado.
              schema:
                type: string
                example: /expenses/67dd91d7e4b0d92f2f123456
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseSaveResponseDto'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /poc/personal-finances/v1/expenses/{expenseId}:
    put:
      tags:
        - Expenses
      summary: Actualizar gasto
      description: Actualizar un gasto existente identificado por su id.
      operationId: updateExpense
      parameters:
        - $ref: '#/components/parameters/CallerNameHeader'
        - $ref: '#/components/parameters/TraceParentHeader'
        - $ref: '#/components/parameters/UserCodeHeader'
        - $ref: '#/components/parameters/ExpenseIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseUpdateRequestDto'
      responses:
        '204':
          description: No Content
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Expenses
      summary: Eliminar gasto
      description: Eliminar gasto existente identificado por su id.
      operationId: deleteExpense
      parameters:
        - $ref: '#/components/parameters/CallerNameHeader'
        - $ref: '#/components/parameters/TraceParentHeader'
        - $ref: '#/components/parameters/UserCodeHeader'
        - $ref: '#/components/parameters/ExpenseIdPath'
      responses:
        '204':
          description: No Content
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /poc/personal-finances/v1/expenses/refresh:
    post:
      tags:
        - Expenses
      summary: Refrescar gastos del usuario
      description: |
        Ejecuta un proceso de refresco de gastos para el correo electrónico indicado.
      operationId: refreshExpenses
      parameters:
        - $ref: '#/components/parameters/CallerNameHeader'
        - $ref: '#/components/parameters/TraceParentHeader'
        - $ref: '#/components/parameters/UserCodeHeader'
      responses:
        '204':
          description: No Content
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /poc/personal-finances/v1/expenses/csv/import:
    post:
      tags:
        - CSV Expenses
      summary: Importar gastos desde CSV
      description: |
        Importa gastos para un usuario a partir de un archivo CSV enviado en multipart/form-data.
      operationId: importCsv
      parameters:
        - $ref: '#/components/parameters/CallerNameHeader'
        - $ref: '#/components/parameters/TraceParentHeader'
        - $ref: '#/components/parameters/UserCodeHeader'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ImportExpenseCsvRequest'
      responses:
        '201':
          description: Archivo CSV importado correctamente.
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /poc/personal-finances/v1/expenses/csv/export:
    get:
      tags:
        - CSV Expenses
      summary: Exportar gastos a CSV
      description: |
        Exporta los gastos del usuario en formato CSV aplicando filtros opcionales.
      operationId: exportCsv
      parameters:
        - $ref: '#/components/parameters/CallerNameHeader'
        - $ref: '#/components/parameters/TraceParentHeader'
        - $ref: '#/components/parameters/UserCodeHeader'
        - $ref: '#/components/parameters/RecipientQuery'
        - $ref: '#/components/parameters/CategoryQuery'
        - $ref: '#/components/parameters/CurrencyQuery'
        - $ref: '#/components/parameters/FromDateQuery'
        - $ref: '#/components/parameters/ToDateQuery'
      responses:
        '200':
          description: Archivo CSV generado correctamente.
          headers:
            Content-Disposition:
              description: Nombre sugerido del archivo descargable.
              schema:
                type: string
                example: attachment; filename="expenses_20260319_201530.csv"
            Content-Type:
              description: Tipo de contenido devuelto por el endpoint.
              schema:
                type: string
                example: text/csv; charset=UTF-8
          content:
            text/csv:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  parameters:
    CallerNameHeader:
      name: caller-name
      in: header
      required: true
      schema:
        $ref: '#/components/schemas/CallerNameHeaderSchema'

    TraceParentHeader:
      name: trace-parent
      in: header
      required: true
      schema:
        $ref: '#/components/schemas/TraceParentHeaderSchema'

    UserCodeHeader:
      name: user-code
      in: header
      required: true
      schema:
        $ref: '#/components/schemas/UserCodeHeaderSchema'

    ExpenseIdPath:
      name: expenseId
      in: path
      required: true
      description: Identificador único del gasto.
      schema:
        $ref: '#/components/schemas/ExpenseIdSchema'

    RecipientQuery:
      name: recipient
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/RecipientSchema'

    CategoryQuery:
      name: category
      in: query
      required: false
      description: Categoría del gasto.
      schema:
        $ref: '#/components/schemas/CategorySchema'

    CurrencyQuery:
      name: currency
      in: query
      required: false
      description: Moneda del gasto.
      schema:
        $ref: '#/components/schemas/CurrencyCodeSchema'

    FromDateQuery:
      name: from
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/DateIsoSchema'

    ToDateQuery:
      name: to
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/DateIsoSchema'

    EncodedCursorQuery:
      name: encodedCursor
      in: query
      required: false
      description: Cursor codificado para obtener la siguiente página de resultados.
      schema:
        $ref: '#/components/schemas/EncodedCursorSchema'

  responses:
    BadRequestError:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDto'

    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDto'

    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDto'

    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDto'

  schemas:
    CallerNameHeaderSchema:
      type: string
      description: Componente que realizó la invocación al servicio.
      pattern: '^[a-zA-Z0-9._-]+$'
      example: web

    TraceParentHeaderSchema:
      type: string
      description: Cabecera de trazabilidad distribuida.
      pattern: '^[0-9A-Fa-f]{2}-[0-9A-Fa-f]{32}-[0-9A-Fa-f]{16}-[0-9A-Fa-f]{2}$'
      example: 00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-00

    ExpenseIdSchema:
      type: string
      description: Identificador único del gasto.
      example: 67dd91d7e4b0d92f2f123456

    GmailMessageIdSchema:
      type: string
      description: Identificador del mensaje de Gmail asociado al gasto, si existe.
      example: 19cf42db92ad643c

    UserCodeHeaderSchema:
      type: string
      format: email
      description: Identificador del usuario en formato de correo electrónico.
      example: user@example.com

    RecipientSchema:
      type: string
      description: Texto contenido en destinatario del gasto.
      example: CIBERTEC

    CommentsSchema:
      type: string
      description: Comentario asociado al gasto.
      example: Mensualidad junio

    CategorySchema:
      type: string
      description: Categoría principal del gasto.
      example: ESTUDIOS

    SourceSchema:
      type: string
      description: Origen o medio desde el cual se generó el gasto.
      example: BBVA_PLIN

    EncodedCursorSchema:
      type: string
      description: Cursor codificado para continuar la paginación.
      example: eyJpZCI6IjY3ZGQ5MWQ3ZTRiMGQ5MmYyZjEyMzQ1NiJ9

    DateIsoSchema:
      type: string
      description: Fecha en formato ISO-8601.
      format: date-time
      example: 2026-03-28T20:25:00Z

    CurrencyCodeSchema:
      type: string
      description: Código de moneda del gasto.
      pattern: '^(PEN|USD)$'
      example: PEN

    PositiveAmountSchema:
      type: number
      description: Monto del gasto. Debe ser mayor a cero.
      format: bigdecimal
      exclusiveMinimum: true
      minimum: 0
      example: 25.50

    ExpenseSaveRequestDto:
      type: object
      description: Datos requeridos para registrar un gasto.
      required:
        - userEmail
        - date
        - detail
        - currency
        - amount
      properties:
        userEmail:
          $ref: '#/components/schemas/UserCodeHeaderSchema'
        gmailMessageId:
          $ref: '#/components/schemas/GmailMessageIdSchema'
        date:
          $ref: '#/components/schemas/DateIsoSchema'
        detail:
          $ref: '#/components/schemas/ExpenseSaveDetail'
        currency:
          $ref: '#/components/schemas/CurrencyCodeSchema'
        amount:
          $ref: '#/components/schemas/PositiveAmountSchema'

    ExpenseSaveDetail:
      type: object
      description: Detalle del gasto a registrar.
      required:
        - source
      properties:
        category:
          $ref: '#/components/schemas/CategorySchema'
        comments:
          $ref: '#/components/schemas/CommentsSchema'
        recipient:
          $ref: '#/components/schemas/RecipientSchema'
        source:
          $ref: '#/components/schemas/SourceSchema'

    ExpenseUpdateRequestDto:
      type: object
      description: Datos requeridos para actualizar un gasto.
      required:
        - date
        - detail
        - currency
        - amount
      properties:
        date:
          $ref: '#/components/schemas/DateIsoSchema'
        detail:
          $ref: '#/components/schemas/ExpenseUpdateDetail'
        currency:
          $ref: '#/components/schemas/CurrencyCodeSchema'
        amount:
          $ref: '#/components/schemas/PositiveAmountSchema'

    ExpenseUpdateDetail:
      type: object
      description: Detalle del gasto a actualizar.
      required:
        - category
        - comments
      properties:
        category:
          $ref: '#/components/schemas/CategorySchema'
        comments:
          $ref: '#/components/schemas/CommentsSchema'
        recipient:
          $ref: '#/components/schemas/RecipientSchema'

    ExpenseSaveResponseDto:
      type: object
      description: Respuesta al registrar un gasto.
      required:
        - expenseId
      properties:
        expenseId:
          $ref: '#/components/schemas/ExpenseIdSchema'

    ExpenseSearchResponseDto:
      type: object
      description: Resultado paginado de la búsqueda de gastos.
      properties:
        expenses:
          type: array
          description: Lista de gastos encontrados.
          items:
            $ref: '#/components/schemas/ExpenseResponseDto'
        nextCursor:
          $ref: '#/components/schemas/EncodedCursorSchema'

    ExpenseResponseDto:
      type: object
      description: Representación de un gasto.
      properties:
        id:
          $ref: '#/components/schemas/ExpenseIdSchema'
        gmailMessageId:
          $ref: '#/components/schemas/GmailMessageIdSchema'
        date:
          $ref: '#/components/schemas/ExpenseResponseDate'
        detail:
          $ref: '#/components/schemas/ExpenseResponseDetail'
        currency:
          $ref: '#/components/schemas/ExpenseResponseCurrency'
        amount:
          $ref: '#/components/schemas/PositiveAmountSchema'

    ExpenseResponseCurrency:
      type: object
      description: Datos de la moneda del gasto.
      properties:
        code:
          type: string
          description: Código de la moneda.
          example: PEN
        symbol:
          type: string
          description: Símbolo de la moneda.
          example: S/

    ExpenseResponseDate:
      type: object
      description: Fecha del gasto en formatos técnico y legible.
      properties:
        utc:
          type: string
          format: date-time
          description: Fecha y hora en UTC.
          example: 2026-03-19T15:30:00Z
        formatted:
          type: string
          description: Fecha formateada para visualización.
          example: 19 Mar 2026

    ExpenseResponseDetail:
      type: object
      description: Detalle del gasto.
      properties:
        category:
          $ref: '#/components/schemas/CategorySchema'
        comments:
          $ref: '#/components/schemas/CommentsSchema'
        recipient:
          $ref: '#/components/schemas/RecipientSchema'
        source:
          $ref: '#/components/schemas/SourceSchema'

    ImportExpenseCsvRequest:
      type: object
      description: Archivo CSV a importar.
      required:
        - file
      properties:
        file:
          type: string
          format: binary
          description: Archivo CSV con los gastos a importar.

    ErrorDto:
      type: object
      description: Estructura estándar de error de la API.
      properties:
        origin:
          $ref: '#/components/schemas/ErrorOrigin'
        code:
          type: string
          description: Código funcional o técnico del error.
          example: Default
        message:
          type: string
          description: Descripción del error.
          example: No hemos podido realizar tu operación. Estamos trabajando para solucionar el inconveniente.

    ErrorOrigin:
      type: string
      description: Origen del error.
      enum:
        - OWN
        - PARTNER
        - THIRD_PARTY
      example: OWN